move server core into separate folder
