+++ onnxruntime/hosting/http/core/http_server.h
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#ifndef ONNXRUNTIME_HOSTING_HTTP_BEAST_HTTP_H
-#define ONNXRUNTIME_HOSTING_HTTP_BEAST_HTTP_H
+#ifndef ONNXRUNTIME_HOSTING_HTTP_CORE_HTTP_SERVER_H
+#define ONNXRUNTIME_HOSTING_HTTP_CORE_HTTP_SERVER_H
 
 #include <functional>
 #include <iostream>
