+++ onnxruntime/hosting/http/core/routes.h
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-#ifndef ONNXRUNTIME_HOSTING_HTTP_ROUTES_H
-#define ONNXRUNTIME_HOSTING_HTTP_ROUTES_H
+#ifndef ONNXRUNTIME_HOSTING_HTTP_CORE_ROUTES_H
+#define ONNXRUNTIME_HOSTING_HTTP_CORE_ROUTES_H
 
 #include <boost/beast/http.hpp>
